[BACKLOG-38516]-code changes for moving Mail (step) from core to plugin
