+++ engine/src/main/resources/kettle-steps.xml
@@ -69,7 +69,6 @@
  <step id="RandomValue"> <description>i18n:org.pentaho.di.trans.step:BaseStep.TypeLongDesc.RandomValue</description> <classname>org.pentaho.di.trans.steps.randomvalue.RandomValueMeta</classname> <category>i18n:org.pentaho.di.trans.step:BaseStep.Category.Input</category> <tooltip>i18n:org.pentaho.di.trans.step:BaseStep.TypeTooltipDesc.RandomValue</tooltip> <iconfile>ui/images/RVA.svg</iconfile> <documentation_url>http://wiki.pentaho.com/display/EAI/Generate+Random+Value</documentation_url> <cases_url/> <forum_url/> </step>
  <step id="GetSubFolders"> <description>i18n:org.pentaho.di.trans.step:BaseStep.TypeLongDesc.GetSubFolders</description> <classname>org.pentaho.di.trans.steps.getsubfolders.GetSubFoldersMeta</classname> <category>i18n:org.pentaho.di.trans.step:BaseStep.Category.Input</category> <tooltip>i18n:org.pentaho.di.trans.step:BaseStep.TypeTooltipDesc.GetSubFolders</tooltip> <iconfile>ui/images/LSF.svg</iconfile> <documentation_url>http://wiki.pentaho.com/display/EAI/Get+SubFolder+names</documentation_url> <cases_url/> <forum_url/> </step>
  <step id="MailValidator"> <description>i18n:org.pentaho.di.trans.step:BaseStep.TypeLongDesc.MailValidator</description> <classname>org.pentaho.di.trans.steps.mailvalidator.MailValidatorMeta</classname> <category>i18n:org.pentaho.di.trans.step:BaseStep.Category.Validation</category> <tooltip>i18n:org.pentaho.di.trans.step:BaseStep.TypeTooltipDesc.MailValidator</tooltip> <iconfile>ui/images/MAV.svg</iconfile> <documentation_url>http://wiki.pentaho.com/display/EAI/Mail+Validator</documentation_url> <cases_url/> <forum_url/> </step>
- <step id="Mail"> <description>i18n:org.pentaho.di.trans.step:BaseStep.TypeLongDesc.Mail</description> <classname>org.pentaho.di.trans.steps.mail.MailMeta</classname> <category>i18n:org.pentaho.di.trans.step:BaseStep.Category.Utility</category> <tooltip>i18n:org.pentaho.di.trans.step:BaseStep.TypeTooltipDesc.Mail</tooltip> <iconfile>ui/images/MAIL.svg</iconfile> <documentation_url>http://wiki.pentaho.com/display/EAI/Mail+%28step%29</documentation_url> <cases_url/> <forum_url/> </step>
  <step id="ProcessFiles"> <description>i18n:org.pentaho.di.trans.step:BaseStep.TypeLongDesc.ProcessFiles</description> <classname>org.pentaho.di.trans.steps.processfiles.ProcessFilesMeta</classname> <category>i18n:org.pentaho.di.trans.step:BaseStep.Category.Utility</category> <tooltip>i18n:org.pentaho.di.trans.step:BaseStep.TypeTooltipDesc.ProcessFiles</tooltip> <iconfile>ui/images/PPF.svg</iconfile> <documentation_url>http://wiki.pentaho.com/display/EAI/Process+files</documentation_url> <cases_url/> <forum_url/> </step>
  <step id="DetectLastRow"> <description>i18n:org.pentaho.di.trans.step:BaseStep.TypeLongDesc.DetectLastRow</description> <classname>org.pentaho.di.trans.steps.detectlastrow.DetectLastRowMeta</classname> <category>i18n:org.pentaho.di.trans.step:BaseStep.Category.Flow</category> <tooltip>i18n:org.pentaho.di.trans.step:BaseStep.TypeTooltipDesc.DetectLastRow</tooltip> <iconfile>ui/images/DLR.svg</iconfile> <documentation_url>http://wiki.pentaho.com/display/EAI/Identify+last+row+in+a+stream</documentation_url> <cases_url/> <forum_url/> </step>
  <step id="DetectEmptyStream"> <description>i18n:org.pentaho.di.trans.step:BaseStep.TypeLongDesc.DetectEmptyStream</description> <classname>org.pentaho.di.trans.steps.detectemptystream.DetectEmptyStreamMeta</classname> <category>i18n:org.pentaho.di.trans.step:BaseStep.Category.Flow</category> <tooltip>i18n:org.pentaho.di.trans.step:BaseStep.TypeTooltipDesc.DetectEmptyStream</tooltip> <iconfile>ui/images/EMS.svg</iconfile> <documentation_url>http://wiki.pentaho.com/display/EAI/Detect+empty+stream</documentation_url> <cases_url/> <forum_url/> </step>
