+++ engine/src/main/resources/org/pentaho/di/trans/step/messages/messages_en_US.properties
@@ -126,10 +126,8 @@ BaseStep.ColumnName.EndDate=end_date
 BaseStep.Category.Bulk=Bulk loading
 BaseStep.TypeLongDesc.XBaseInput=XBase input
 BaseStep.status.Stopped=Stopped
-BaseStep.TypeLongDesc.Mail=Mail
 BaseStep.TypeLongDesc.AddConstants=Add constants
 BaseStep.TypeLongDesc.WriteToLog=Write to log
-BaseStep.TypeTooltipDesc.Mail=Send eMail.
 BaseStep.TypeLongDesc.GPBulkLoader=Greenplum bulk loader
 BaseStep.Log.StopPuttingARow=Stopped while putting a row on the buffer
 BaseStepDialog.ErrorHandling.NrErrField.Label=Nr of errors fieldname
