+++ plugins/mail/impl/src/main/resources/org/pentaho/di/trans/steps/mail/messages/messages_en_US.properties
@@ -2,6 +2,8 @@
 
 
 #Mon Jan 04 18:35:29 CET 2010
+BaseStep.TypeLongDesc.Mail=Mail
+BaseStep.TypeTooltipDesc.Mail=Send eMail.
 MailDialog.ReplyToAddresses.Label=Reply to
 Mail.Exception.CouldnotFindServerField=We can not find server field [{0}] in input stream\!
 MailDialog.ZipFilename.Label=Zip filename
