+++ plugins/mail/impl/src/main/resources/org/pentaho/di/trans/steps/mail/messages/messages_ko_KR.properties
