+++ plugins/mail/impl/src/main/resources/org/pentaho/di/trans/steps/mail/messages/messages_zh_CN.properties
